Seperate raw route into it's own file.
