+++ src/main.py
@@ -1,10 +1,8 @@
 from fastapi import FastAPI, APIRouter
 from starlette.middleware.cors import CORSMiddleware
 
-from src.data import api
-from src.routes import base, admin
 
-api = api.API()
+from src.routes import base, admin, raw
 
 app = FastAPI(
     title="HD2 Community API",
@@ -27,19 +25,7 @@ app.add_middleware(
 router = APIRouter()
 
 
-@router.get("/raw/all", status_code=200)
-async def get_all_raw(source: str = ""):
-    """
-    All AHGS API Calls Only
-    """
-    update = await api.update_all()
-    data = {}
-    for i, (k, v) in enumerate(api.raw_data.items()):
-        if v["data"] is not [] or {}:
-            data[k] = v["data"]
-    return data
-
-
 app.include_router(router)
 app.include_router(base.router)
 app.include_router(admin.router)
+app.include_router(raw.router)
